+++ src/bin/check.rs
@@ -1,9 +1,8 @@
 use std::env;
 
-use cargo::ops::{self};
-use cargo::ops::cargo_check::{Options, with_check_env};
+use cargo::ops::{self, CompileOptions, MessageFormat};
+use cargo::ops::with_check_ws;
 use cargo::util::{CliResult, Config};
-
 
 pub const USAGE: &'static str = "
 Check a local package and all of its dependencies for errors
@@ -43,12 +42,61 @@ the manifest. The default profile for this command is `dev`, but passing
 the --release flag will use the `release` profile instead.
 ";
 
+#[derive(RustcDecodable)]
+pub struct Options {
+    flag_package: Vec<String>,
+    flag_jobs: Option<u32>,
+    flag_features: Vec<String>,
+    flag_all_features: bool,
+    flag_no_default_features: bool,
+    flag_target: Option<String>,
+    flag_manifest_path: Option<String>,
+    flag_verbose: u32,
+    flag_quiet: Option<bool>,
+    flag_color: Option<String>,
+    flag_message_format: MessageFormat,
+    flag_release: bool,
+    flag_lib: bool,
+    flag_bin: Vec<String>,
+    flag_example: Vec<String>,
+    flag_test: Vec<String>,
+    flag_bench: Vec<String>,
+    flag_locked: bool,
+    flag_frozen: bool,
+}
+
 pub fn execute(options: Options, config: &Config) -> CliResult<Option<()>> {
     debug!("executing; cmd=cargo-check; args={:?}",
            env::args().collect::<Vec<_>>());
 
-    with_check_env(options, config, |ws, opts| {
-        ops::compile(ws, opts)?;
+    config.configure(options.flag_verbose,
+                     options.flag_quiet,
+                     &options.flag_color,
+                     options.flag_frozen,
+                     options.flag_locked)?;
+
+    with_check_ws(options.flag_manifest_path.clone(), config, |ws| {
+        let opts = CompileOptions {
+            config: config,
+            jobs: options.flag_jobs,
+            target: options.flag_target.as_ref().map(|t| &t[..]),
+            features: &options.flag_features,
+            all_features: options.flag_all_features,
+            no_default_features: options.flag_no_default_features,
+            spec: ops::Packages::Packages(&options.flag_package),
+            mode: ops::CompileMode::Check,
+            release: options.flag_release,
+            filter: ops::CompileFilter::new(options.flag_lib,
+                                            &options.flag_bin,
+                                            &options.flag_test,
+                                            &options.flag_example,
+                                            &options.flag_bench),
+            message_format: options.flag_message_format,
+            target_rustdoc_args: None,
+            target_rustc_args: None,
+        };
+
+        ops::compile(ws, &opts)?;
         Ok(None)
     })
 }
